Refactor vis tests and finish coverage
